CUT: width * height was computed in int and overflowed it
    unsigned size = header.width * header.height;

Both fields are WORD, so both promote to int and the product is evaluated
there: 65535 * 65535 is 4294836225, well past INT_MAX. That is undefined
behaviour, and while two's-complement wrap happens to give the right value
back once it is converted to unsigned, nothing guarantees it - at -O2 the
compiler is entitled to assume it cannot happen.

Multiplying as unsigned keeps the whole range in a type that can hold it.

UBSan reported "signed integer overflow: 65520 * 65535 cannot be represented
in type 'int'" on a header declaring those dimensions; it no longer does, and
the file is refused when the allocation fails, as before.

Found by the mutation fuzzer. Finding 2 of AUDIT.md.

+++ Source/FreeImage/PluginCUT.cpp
@@ -156,7 +156,10 @@ Load(FreeImageIO *io, fi_handle handle, int page, int flags, void *data) {
 
 		unsigned i = 0, k = 0;
 		unsigned pitch = FreeImage_GetPitch(dib);
-		unsigned size = header.width * header.height;
+		// both are WORD, so each promotes to int and the product is computed in
+		// int: 65535 * 65535 overflows it.  Multiply as unsigned, where the full
+		// range fits.
+		unsigned size = (unsigned)header.width * (unsigned)header.height;
 		BYTE count = 0, run = 0;
 
 		while (i < size) {
